+++ CHANGELOG.md
@@ -17,3 +17,10 @@
 ## V1.0.2
 - Added catalogues and roblox_cores apps
 - Added Models to catalogues and roblox_cores apps
+
+## V1.0.3
+- Added serializers.py in apps.roblox_cores.serializers 
+- Added urls.py in apps.roblox_cores.urls
+- Added get, post, put, and delete in apps.roblox_cores.views
+- Added apps.roblox_cores.admin for Django Admin
+- Minor changes in apps.roblox_cores.models